feat: ユーザーの学習時間を管理するロジックを追加し、関連する関数を実装
